Updated GetValue to return a ulong to fit with Game Time package

+++ Public/ITimeSolver.cs
@@ -2,6 +2,6 @@ namespace FedoraDev.NPCSchedule
 {
     public interface ITimeSolver
     {
-        float GetValue();
+        ulong GetValue();
     }
 }
